requirements.txt
langchain
langchain_community
langchain-community
langchain-openai
openai
python-dotenv
streamlit